fixed singleton pattern, closes #15
due to passing contexts that are not meant to be used with toast,
many times when the app would like to show a dialog, an exception is
thrown stating that the looper in the context has not been initialized.

By reworking the singleton pattern that has not been fully utilized, all
calls will now try to go through PhotosynqApplication.sApplication which
is a valid context for toast messages.